Update vm.cpp
oh no! so I accidentally committed to main on my last attempt. I brought the code back though so you'll see the knew loop Katia made here but it might not be in the main branch sorry!

So this request should include here code she just had the PR on and then my new code with the switch statements for control operations.